% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{Function that accesses files in the current package}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Character vectors, specifying subdirectory and file(s) within the package.}
}
\description{
Function that accesses files in the current package
}
