% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_.R
\name{body_}
\alias{body_}
\title{Body}
\usage{
body_(top_packages, today, last_month)
}
\arguments{
\item{top_packages}{A character vector of packages names.}

\item{today}{Today's date.}

\item{last_month}{Last month's date.}
}
\description{
Body
}
