% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_server.R
\name{python_plot}
\alias{python_plot}
\title{Server for Python tab}
\usage{
python_plot(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The x variable.}

\item{y}{The y variable.}
}
\description{
Server for Python tab
}
