% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_server.R
\name{r_plot_hc}
\alias{r_plot_hc}
\title{Server for R tab (Highcharter)}
\usage{
r_plot_hc(data, x, y, group)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The x variable.}

\item{y}{The y variable.}

\item{group}{Grouping variable}
}
\description{
Server for R tab (Highcharter)
}
