% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-pool-bigrquery.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dbPool}
\alias{poolClose}
\alias{bq_auth}
\alias{bigquery}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigrquery}{\code{\link[bigrquery]{bigquery}}, \code{\link[bigrquery]{bq_auth}}}

  \item{pool}{\code{\link[pool]{dbPool}}, \code{\link[pool:Pool-class]{poolClose}}}
}}

