% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Application}
\usage{
run_app(
  top_packages = cran_top_downloads(count = 5)$package,
  today = Sys.Date() - 1,
  last_month = Sys.Date() - 61
)
}
\arguments{
\item{top_packages}{A character vector of packages names.}

\item{today}{Today's date.}

\item{last_month}{Last month's date.}
}
\description{
Application
}
