% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run application}
\usage{
runApp(top_packages, today, last_month)
}
\arguments{
\item{top_packages}{A character vector of packages names.}

\item{today}{Today's date.}

\item{last_month}{Last month's date.}
}
\description{
Run application
}
